challenge: Extract ChallengeInfo, add MergeContext
Defines a new MergeContext type which will be used to store information
about an ongoing merge. Additionally, extracts the ChallengeInfo type to
this file and adds mode as that will be needed by some merge steps.
