+++ challenge-server/api.ts
@@ -437,7 +437,7 @@ async function mergeTestEvents(req: Request, res: Response): Promise<void> {
     }
 
     const tracer = new MergeTracer();
-    const merger = new Merger(stage, clients);
+    const merger = new Merger(mergeData.challengeInfo, stage, clients);
     const result = merger.merge(tracer);
     if (result === null) {
       res.status(500).send();
