+++ challenge-server/merging/client-events.ts
@@ -12,15 +12,15 @@ import {
 import { ChallengeEvents } from '@blert/common/generated/challenge_storage_pb';
 import { Event } from '@blert/common/generated/event_pb';
 
-import logger from '../log';
-import { ChallengeInfo } from './merge';
-import { PlayerState, TickState, TickStateArray } from './tick-state';
 import {
   consistencyCheckerForStage,
   ConsistencyIssue,
   ConsistencyIssueType,
   MovementConsistencyChecker,
 } from './consistency';
+import { ChallengeInfo } from './context';
+import logger from '../log';
+import { PlayerState, TickState, TickStateArray } from './tick-state';
 
 export const enum ClientAnomaly {
   MULTIPLE_PRIMARY_PLAYERS = 'MULTIPLE_PRIMARY_PLAYERS',
